requirements.txt
numpy~=1.23
pillow~=9.5
click~=8.1
PyYAML~=6.0
fastapi~=0.115.5
uvicorn~=0.32.1
structlog~=24.1.0
omegaconf~=2.3.0
psycopg2-binary~=2.9.10
alembic~=1.15.1
passlib~=1.7.4
bcrypt~=4.3.0
jinja2~=3.1.6
python-multipart~=0.0.20
pydantic[email]
pydantic~=2.11.1
python-dotenv~=1.1.0
aiofiles~=24.1.0
pytest==8.3.5
pytest-cov~=6.1.1
pytest-asyncio==0.26.0
locust~=2.36.2
scikit-learn~=1.6.1
